Made sure android native app is ready to be working on more. Updated headers, and libs for unity game project.
Android native app both bootstrap and native one now implemented with multiple of activities (screens) and ready for android team to use as a base to work more on it.
Remember that CustomNativeActivity or any class that extend UnityPlayerNativeActivity can only be switched to or created when integrated has been done, not for bootstrap.

In this case, I created another activity so bootstrap has more meaningful way to test functionality in nativeprojectlib.

So UnityPlayerNativeActivity class will be only used when integration has been done (Unity + Android), other than that bootstrap project can test other activity classes.